Fix typization in is-my-friend-cheating.cpp

+++ is-my-friend-cheating.cpp
@@ -4,11 +4,11 @@ using namespace std;
 
 class RemovedNumbers {
   public:
-    static vector<vector<long long>> removNb(long n) {
+    static vector<vector<long long>> removNb(long long n) {
       vector<vector<long long>> result;
-      long sequence_sum = n * (n + 1) / 2;
-      for (long x = 1; x <= n; x++) {
-        long y = (sequence_sum - x) / (x + 1);
+      long long sequence_sum = n * (n + 1) / 2;
+      for (long long x = 1; x <= n; x++) {
+        long long y = (sequence_sum - x) / (x + 1);
         if (y <= n && x * y == sequence_sum - x - y) {
           result.push_back({x, y});
         }
